+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* Add KohanaPSRLogger to provide a standard PSR/Log interface to the Kohana log
+
 ## v0.3.4 (2018-09-06)
 
 * Define class_alias for PHPUnit_Framework_Constraint -> namespaced version in